requirements.txt
beautifulsoup4
lxml
python-dotenv
httpx[http2]
motor
pybloom-live
playwright
tenacity
simhash-py
simhash